renamed cardContainer to scrollContainer
